fix(meter_code): require dashed meter codes across MQTT and ESPHome
switch meter code parsing/validation to dashed formats: YY-serial and optional YY-serial-NNN
update standalone firmware parsers/config providers and ESPHome validator logic
migrate all examples and user docs to dashed meter code input
update CI placeholder/injected dummy values to dashed METER_CODE format
refresh ESPHOME-release artifacts after source updates
add/adjust parser edge-case tests for dashed format and invalid inputs

+++ ESPHOME-release/everblu_meter/everblu_meter.cpp
@@ -353,7 +353,7 @@ namespace esphome
         void EverbluMeterComponent::dump_config()
         {
             ESP_LOGCONFIG(TAG, "EverBlu Meter:");
-            ESP_LOGCONFIG(TAG, "  Meter Code: %s (year=%u, serial=%u)", meter_code_.c_str(), meter_year_, meter_serial_);
+            ESP_LOGCONFIG(TAG, "  Meter Code: %s (year=%u, serial=%lu)", meter_code_.c_str(), meter_year_, (unsigned long)meter_serial_);
             ESP_LOGCONFIG(TAG, "  Meter Type: %s", is_gas_ ? "Gas" : "Water");
             ESP_LOGCONFIG(TAG, "  Everblu Meter Component Version: %s", EVERBLU_FW_VERSION);
             if (is_gas_)
